Add files via upload
Ludo Game project is written in Python. The project file contains python scripts (game.py, run.py, painter.py, recorder.py). This is a simple console based strategy board game which is very easy to understand and use. Talking about the gameplay, all the playing rules are the same just like we play in real time ludo. Here at first, the user has to select players i.e either human or computer. After selecting human, the player has to enter details such as name and select color(red, green, yellow and blue). the player can also start the game within two players if he/she wants.

After starting the game, a console based ludo board appears, other rules are the same. First, the computer or the player has to roll the dice. The main thing in this console based game is that the player just has to press “Enter Key” to roll the dice. At the top of the board, it displays a dice with the number. The system keeps on rolling until there’s a possible pawn to move. All the game movements are performed automatically. Whenever the player tries to end the game, there’s an option available to save the unfinished game so that he/she can continue it later. A simple console GUI is provided for the easy gameplay. The gameplay design is so simple that user won’t find it difficult to use and understand.

In order to run the project, you must have installed Python, on your PC. This is a simple console Based game written for the beginners. Simple Ludo Game in Python project with source code is free to download. Use for education purpose only! For the project demo, have a look at the YouTube Video above.